+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.2.2-SNAPSHOT
+# 0.3.0
 
 * Replace `@motif.Dependencies` API with `ScopeFactory.create`. See https://github.com/uber/motif/issues/125 for details.
 * Do not use computeIfAbsent in ScopeFactory in order to avoid ConcurrentModificationException in the IntelliJ plugin